Build Windows & Unix `noarch` variants
To capture the additional Windows requirements without loosing the value
of `noarch`, add a variant to build a Unix package and a Windows
package. Include a hard dependency on the OS virtual package to ensure
the Unix variant isn't installed on Windows and vice versa. This in
addition to the other dependencies a platform may require.